Renamed some folders + moved EmuShared.h
