+++ src/common/win32/EmuShared.h
