+++ src/common/win32/InlineFunc.cpp
