+++ src/core/HLE/XACTENG/XactEng.cpp
@@ -47,7 +47,7 @@ namespace xboxkrnl
 #include "Common/Logging.h"
 #include "CxbxKrnl/Emu.h"
 #include "CxbxKrnl/EmuFS.h"
-#include "CxbxKrnl/EmuShared.h"
+#include "EmuShared.h"
 #include "CxbxKrnl/EmuXTL.h"
 #include "CxbxKrnl/VMManager.h"
