+++ src/core/HLE/XAPI/OHCI/XInput/DInput.cpp
@@ -39,7 +39,7 @@
 
 #include "CxbxKrnl/Emu.h"
 #include "CxbxKrnl/EmuXTL.h"
-#include "CxbxKrnl/EmuShared.h"
+#include "EmuShared.h"
 #include "Common/Win32/DInputController.h"
 #include "Logging.h"
